[helm] Fix broken `helm init` for V2 + add test
The main fix here is adding the HELM_VERSION ENV into the Dockerfile.
The ARG that was already specified is only available at Docker image
build-time, so checking for Helm v2 in the entrypoint script didn't work
properly and never ran a `helm init`

I also added a test step that should catch this in the future (and is
actually specifed as a Contribution Requirement for this repo), plus
made the entrypoint script fail in case of unexpected errors (with the
bash '-e' param).

The new SKIP_CLUSTER_CONFIG env var is used for testing, but also handy
for users that want to run Helm commands that don't need any of that,
for example `helm template`.

I also updated the `gcloud container ...` commands to match the latest
interface. Beta is no longer required.