Fixes trivy image so it works in google cloud build (#437)
After this commit,
https://github.com/aquasecurity/trivy/commit/9c91da8a2b8619c5617d72c31031324121525af4#diff-3254677a7917c6c01f55212f86c57fbf,
the trivy image will not have access to the docker socket and will not be able
to write any scan output due to permissions issues.  Another way I thought of
to fix this was to change the user in the build step, but I don't see any
documentation for how to do this in google cloud build.