Update setup documentation for Firebase build image
While trying to determine why cloud function that I was working with in
the build were not being deployed I found an open issue.
https://github.com/firebase/firebase-tools/issues/4757

This change is to update the documentation to reflect the need for
not just the Firebase Admin role in the builds, but also the
Service Accounts role.

Thanks to @akauppi for discovering the issue. It was the same issue
I had spent multiple hours trying to figure out.