firebase README: Add note for Windows users to convert line endings in firebase.bash to LF
This update adds a section to the README that instructs Windows users to run the `dos2unix` command on the `firebase.bash` file to convert line endings from CRLF to LF. 

On Windows machines, the default line endings are CRLF (Carriage Return Line Feed), while Linux-based environments (such as Docker) expect LF (Line Feed) line endings. If the `firebase.bash` file is not converted to LF, it may fail to execute properly, leading to errors like "no such file or directory" during the build process.

By adding this instruction, we ensure that users on Windows machines are aware of the potential issue and can resolve it before submitting or building the Docker image.